[DD-124] security 수정
