+++ src/main/java/shop/dalda/config/SecurityConfig.java
@@ -44,7 +44,8 @@ public class SecurityConfig {
             "/configuration/ui",
             "/configuration/security",
             "/swagger-ui.html",
-            "/webjars/**"
+            "/webjars/**",
+            "/api/home/**"
     };
 
     @Bean
